Added ability to check for Empty Instructions.
Added a shutdown in the event that we overrun program memory, this will generate
a number of empty instructions. Once a threshold is hit, it will shutdown the system.

Removed the legacy system, the new system structure is up to the same spot as the
the legacy system, so it is okay to remove that part of the code base.