Started the implementation of OpCodes
Reworked how attaching is done so that each piece has access to the system in general.
This is suboptimal but I would really like to get this thing functioning so I opened up access
to all system resources for every component.

Updated the handling of OpCodes so that they can execute themselves when they are
told to by the CPU.